+++ CHANGELOG.rst
@@ -2,6 +2,18 @@
 CHANGELOG
 =========
 
+Version 1.3.2
+-------------
+
+* fasta added --disambiguate option to turn ambiguous sequences into
+  all permutations possible
+
+Version 1.3.1
+-------------
+
+* plot_muts added --cluster and --html options
+* fasta added --split and --wrap options
+
 Version 1.3.0
 -------------